feat: 보충 블록 — 본문을 고치지 않는 개인화 (P4.3)
막힌 지점 뒤에 설명을 끼운다. 본문 블록을 바꾸는 경로가 테이블·도메인·합성 함수 세 층에
모두 없다 — 규칙으로 두면 언젠가 깨지고, 본문이 사용자마다 달라지는 순간 집단 루프가 멈춘다.

- 블록 ID 를 저장하지 않고 행의 PK 에서 만든다(s{pk}). 순번을 계산해 발급하면 지운 뒤 다시
  만들 때 지워진 ID 가 되살아나고, 그러면 두 설명의 신호가 한 블록에 섞인다. BIGSERIAL 은
  되돌아가지 않으므로 재사용이 구조적으로 불가능하다.
- 생성 시점을 정했다(미결정 항목): 요청 시 즉시 생성하고 레이어에 남긴다. 사전 생성은 아무도
  막히지 않은 지점의 설명까지 만들어야 하고 그 비용이 블록 수 × 사용자 수로 늘어난다.
  같은 지점을 다시 요청하면 이미 만든 것을 돌려주므로 새로고침이 LLM 호출이 되지 않고,
  DB 의 UNIQUE 가 그 방어선의 마지막 층이다.
- AI 호출은 게이트웨이를 지난다. 예산 상한 검사가 그 한 곳에만 있으므로 직접 벤더를 부르면
  뼈대당·일일 상한 밖으로 새는 경로가 생긴다. 뼈대 ID 를 달아 그 뼈대의 상한에 넣는다.
- 프롬프트가 "본문을 다시 쓰지 마라"고 명시하고 그 사용자의 질문·오답을 함께 넣는다. 무엇이
  막혔는지가 없으면 이미 본문에 있는 이야기를 반복하는 설명이 나온다.
- 빈 응답은 저장하지 않는다. 저장하면 다시 요청해도 "이미 있다"로 판정돼 그 지점이 영구히
  빈칸이 된다.

M4 DoD 2(보충 삽입, 본문 불변)와 4(버전이 올라도 기록 유지)를 확인했다.

+++ nextchapter-domain/src/main/java/com/mindplates/nextchapter/domain/chapter/model/BlockIds.java
@@ -49,8 +49,13 @@ public final class BlockIds {
         return prefixed(BODY_PREFIX, sequence);
     }
 
-    /** 보충 블록 ID. 공용 본문과 다른 공간이라 신호가 섞이지 않는다. */
-    public static String ofSupplement(int sequence) {
+    /**
+     * 보충 블록 ID. 공용 본문과 다른 공간이라 신호가 섞이지 않는다.
+     *
+     * <p>{@code long} 을 받는 이유는 발급자가 <b>테이블의 PK</b> 이기 때문이다. 순번을 계산해 발급하면 지운
+     * 뒤 다시 만들 때 지워진 ID 가 되살아나고, 그러면 두 설명의 신호가 한 블록에 섞인다.
+     */
+    public static String ofSupplement(long sequence) {
         return prefixed(SUPPLEMENT_PREFIX, sequence);
     }
 
@@ -67,7 +72,7 @@ public final class BlockIds {
         return isValid(blockId) && blockId.startsWith(SUPPLEMENT_PREFIX);
     }
 
-    private static String prefixed(String prefix, int sequence) {
+    private static String prefixed(String prefix, long sequence) {
         if (sequence < 1) {
             throw new IllegalArgumentException("블록 ID 순번은 1 이상이어야 합니다: " + sequence);
         }
